Creating proper package for eodag_sentinelsat
