--- providers.yml
@@ -1,188 +1,0 @@
-
----
-  !provider
-name: scihub
-priority: 0
-api: !plugin
-  type: SentinelsatAPI
-  endpoint: 'https://scihub.copernicus.eu/apihub/'
-  dont_quote:
-    - '['
-    - ']'
-    - '$'
-    - '='
-    - '&'
-    - ':'
-  pagination:
-    next_page_url_tpl: '{url}/odata/v1/Products?{search}&$top={items_per_page}&$skip={skip}'
-    total_items_nb_key_path: 'https://scihub.copernicus.eu/apihub/odata/v1/Products/$count'
-  results_entry: 'value'
-  # literal_search_params:
-  #   $format: json
-  # free_text_search_operations:
-  #   $search:
-  #     union: ' OR '
-  #     wrapper: '"{}"'
-  #     operations:
-  #       AND:
-  #         - 'footprint:"Intersects({geometry#to_wkt})"'
-  #         - '( (platformname:{platform} AND producttype:{productType}) )'
-  #         - '( cloudcoverpercentage:[0 TO {cloudCover}] )'
-  #         - '( beginPosition:[{startTimeFromAscendingNode#to_iso_utc_datetime} TO {completionTimeFromAscendingNode#to_iso_utc_datetime}] AND endPosition:[{startTimeFromAscendingNode#to_iso_utc_datetime} TO {completionTimeFromAscendingNode#to_iso_utc_datetime}] )'
-  #         - '( {raw} )'
-  #         - '({id#remove_extension})'
-  discover_metadata:
-    auto_discovery: true
-    metadata_pattern: '^[a-zA-Z0-9]+$'
-    search_param:
-      free_text_search_operations:
-        $search:
-          operations:
-            AND:
-              - '({metadata}:{{{metadata}}})'
-    metadata_path: '$.*'
-  metadata_mapping:
-    acquisitionType: '$.acquisitiontype'
-    id: '$.uuid'
-    uuid:
-      - uuid
-      - '$.uuid'
-    geometry:
-      - area
-      - '$.footprint'
-    productType:
-      - producttype
-      - '$.producttype'
-    platform:
-      - platformname
-      - '$.platformname'
-    platformSerialIdentifier: '$.platformserialidentifier'
-    instrument: '$.instrumentname'
-    processingLevel: '$.processinglevel'
-    title:
-      - title
-      - '$.title'
-    abstract: '$.summary'
-    orbitNumber:
-      - orbitnumber
-      - '$.orbitnumber'
-    orbitDirection:
-      - orbitdirection
-      - '$.orbitdirection'
-    cloudCover:
-      - cloudcoverpercentage
-      - '$.cloudcoverpercentage'
-    startTimeFromAscendingNode:
-      - start
-      - '$.beginposition'
-    completionTimeFromAscendingNode:
-      - end
-      - '$.endposition'
-    downloadLink: '$.link'
-    polarizationChannels:
-      - polarisationmode
-      - '$.polarisationmode'
-    sensorMode:
-      - sensoroperationalmode
-      - '$.sensoroperationalmode'
-
-    # Additional metadata provided by the providers but that don't appear in the reference spec
-    linkAlternative: '$.link_alternative'
-    ingestionDate: '$.ingestiondate'
-    relativeOrbitNumber:
-      - relativeorbitnumber
-      - '$.relativeorbitnumber'
-    gmlFootprint:
-      - gmlfootprint
-      - '$.gmlfootprint'
-    format:
-      - format
-      - '$.format'
-    processingBaseline: '$.processingbaseline'
-    title:
-      - filename
-      - '$.filename'
-    instrumentShortname: '$.instrumentshortname'
-    size: '$.size'
-    s2DatatakeId: '$.s2datatakeid'
-    platformIdentifier: '$.platformidentifier'
-    identifier:
-      - identifier
-      - '$.identifier'
-    raw:
-      - raw
-      - '$.raw'
-    # storageStatus: must be one of ONLINE, STAGING, OFFLINE
-    storageStatus: '{$.storage_status#get_group_name((?P<ONLINE>True)|(?P<OFFLINE>False))}'
-products:
-  S1_SAR_OCN:
-    product_type: OCN
-    collection: S1
-  S1_SAR_GRD:
-    product_type: GRD
-    collection: S1
-  S1_SAR_SLC:
-    product_type: SLC
-    collection: S1
-  S2_MSI_L1C:
-    collection: S2ST
-    product_type: S2MSI1C
-  S2_MSI_L2A:
-    collection: S2ST
-    product_type: S2MSI2A
-  S3_EFR:
-    product_type: OL_1_EFR___
-    collection: S3
-  S3_ERR:
-    product_type: OL_1_ERR___
-    collection: S3
-  S3_OLCI_L2LFR:
-    product_type: OL_2_LFR___
-    collection: S3
-  S3_OLCI_L2LRR:
-    product_type: OL_2_LRR___
-    collection: S3
-  S3_SLSTR_L1RBT:
-    product_type: SL_1_RBT___
-    collection: S3
-  S3_SLSTR_L2LST:
-    product_type: SL_2_LST___
-    collection: S3
-  S3_LAN:
-    product_type: SR_2_LAN___
-    collection: S3
-
-# --- Available keywords: ---
-#'acquisitiontype': 'NOMINAL',
-#'beginposition': datetime.datetime(2017, 4, 25, 15, 56, 12, 814000), ===> replaced by date
-#'endposition': datetime.datetime(2017, 4, 25, 15, 56, 39, 758000), ===> replaced by date
-#'filename': 'S1A_IW_SLC__1SDV_20170425T155612_20170425T155639_016302_01AF91_46FF.SAFE',
-#'footprint': 'POLYGON ((34.322010 0.401648,36.540989 0.876987,36.884121 -0.747357,34.664474 -1.227940,34.322010 0.401648))', ===> replaced by area
-#'format': 'SAFE',
-#'gmlfootprint': '<gml:Polygon ... >'
-#'identifier': 'S1A_IW_SLC__1SDV_20170425T155612_20170425T155639_016302_01AF91_46FF',
-#'ingestiondate': datetime.datetime(2017, 4, 25, 19, 23, 45, 956000),
-#'instrumentname': 'Synthetic Aperture Radar (C-band)',
-#'instrumentshortname': 'SAR-C SAR',
-#'lastorbitnumber': 16302,
-#'lastrelativeorbitnumber': 130,
-#'link': "https://scihub.copernicus.eu/apihub/odata/v1/Products('04548172-c64a-418f-8e83-7a4d148adf1e')/$value",
-#'link_alternative': "https://scihub.copernicus.eu/apihub/odata/v1/Products('04548172-c64a-418f-8e83-7a4d148adf1e')/",
-#'link_icon': "https://scihub.copernicus.eu/apihub/odata/v1/Products('04548172-c64a-418f-8e83-7a4d148adf1e')/Products('Quicklook')/$value",
-#'missiondatatakeid': 110481,
-#'orbitdirection': 'ASCENDING',
-#'orbitnumber': 16302,
-#'platformidentifier': '2014-016A',
-#'platformname': 'Sentinel-1',
-#'polarisationmode': 'VV VH',
-#'productclass': 'S',
-#'producttype': 'SLC',
-#'relativeorbitnumber': 130,
-#'sensoroperationalmode': 'IW',
-#'size': '7.1 GB',
-#'slicenumber': 8,
-#'status': 'ARCHIVED',
-#'summary': 'Date: 2017-04-25T15:56:12.814Z, Instrument: SAR-C SAR, Mode: VV VH, Satellite: Sentinel-1, Size: 7.1 GB',
-#'swathidentifier': 'IW1 IW2 IW3',
-#'title': 'S1A_IW_SLC__1SDV_20170425T155612_20170425T155639_016302_01AF91_46FF',
-#'uuid': '04548172-c64a-418f-8e83-7a4d148adf1e'
